refactor: separate MJ tests files
